Introduce new config option to skip not existing artifacts
We had the problem that we generate a P2 site, but some of the artifacts
are not there yet. Instead of generating lots of standalone sites we decided
to introduce this skip option. With this option we can generate a site,
and then build the missing artifacts (which need that site). After that we
regenerate the site again to include that artifact.

The slution is to catch any exception in the artifact resolve loop and
only log, if the option is enabled. It would be better to catch the concrete
excpetion at this point. But for unknown reason this checked exception does
not occur in the method signature, so we must catch the base class.