Convert tabs to spaces
Some parts of the README.md were indented with tabs, so I converted
them to four spaces to comply to the rest of the document.

Also dedented the code blocks because for blocks marked up with ```
it is not necessary to indent them with four spaces to render them
as preformated. It should be used either ``` or four spaces to markup
preformatted text.