fix #116: Feature skip not existing artifacts (#122)
* Fix indendation

The whole file uses spaces so I converted tabs to spaces.

* Introduce new config option to skip not existing artifacts

We had the problem that we generate a P2 site, but some of the artifacts
are not there yet. Instead of generating lots of standalone sites we decided
to introduce this skip option. With this option we can generate a site,
and then build the missing artifacts (which need that site). After that we
regenerate the site again to include that artifact.

The slution is to catch any exception in the artifact resolve loop and
only log, if the option is enabled. It would be better to catch the concrete
excpetion at this point. But for unknown reason this checked exception does
not occur in the method signature, so we must catch the base class.

* Remove dangling text in table

* Convert tabs to spaces

Some parts of the README.md were indented with tabs, so I converted
them to four spaces to comply to the rest of the document.

Also dedented the code blocks because for blocks marked up with ```
it is not necessary to indent them with four spaces to render them
as preformated. It should be used either ``` or four spaces to markup
preformatted text.