Fix src/test/integration/feature-14-it (#168)
Remove the dependency on
  org.reficio:p2-maven-plugin:1.0.0-SNAPSHOT
in the test case.

That old SNAPSHOT version was only found in the
  http://repo.reficio.org/maven
repository. That repo does not exist anymore, so the artifact
cannot be used as a test object anymore.

In fact, the assertion on that bundle was removed a while ago in
  validate.groovy
anyways, so it can safely be removed fully.